+++ src/modules/address-book/index.mjs
@@ -1,0 +1,7 @@
+export * from "./category.mjs";
+export * from "./address-book.mjs";
+export * from "./add-to-category.mjs";
+export * from "./remove-from-category.mjs";
+export * from "./create-contact.mjs";
+export * from "./delete-contact.mjs";
+export * from "./update-contact.mjs";
